Complete work for issue #307
Agent: senior_software_engineer
Task: f5ec6a47-e7e7-4f44-96cc-298c92ff2e61

[orchestrator-commit]

(Failsafe commit: auto-staged all changes)

+++ tests/unit/domain/test_workflow.py
@@ -549,7 +549,7 @@ class TestWorkflowQueries:
         duration = workflow.get_duration_seconds()
         assert duration is not None
         assert duration >= 0
-        assert isinstance(duration, int)
+        assert isinstance(duration, float)
 
 
 class TestEventManagement:
